# openapi.yaml
...
openapi: 3.1.0
info:
  title: HALO — Centralised Services control surface
  version: 0.1.0
  summary: Read and change the live configuration of six centralised alerting services.
  description: |-
    HALO is the configuration control surface for the WBGT, Noise, Haze, Lightning, Ailytics and
    Subcon Activities services. Every service reads its configuration from Postgres on a cron, so a
    write through this API changes real behaviour on the next tick — messages sent to real WhatsApp
    groups on real construction sites.

    **There is no staging.** Treat every project as production. Where a project has `enabled = false`
    it is safe to experiment with; anything else is live.

    The column set for each service is not fixed in this document. It is introspected from the live
    database, so new columns appear without a release. Call `getSchema` first and work from what it
    returns, including each field's widget, options, default and help text.

    **Agents:** the same operations are available over MCP at `/api/mcp` (Streamable HTTP, JSON only),
    with tools derived from this document — one tool per `operationId`, carrying safety annotations.
    That is usually the better surface; this one is the contract it is generated from.
  contact:
    name: Wenti Labs
servers:
  - url: http://localhost:5178
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/jobs/{job}:
    post:
      operationId: runJob
      tags:
        - operations
      summary: Trigger a job on an alerting service
      description: |-
        Runs one job against one project for a date range. **Requires the `jobs` scope, which `write` does
        not confer.**

        These have real-world effects that differ per job. `wbgt-scrape` logs into CloudLynx and drives a
        browser session that can run for minutes. The sheet jobs write to Google Sheets. Read the job's
        `description` in `listJobs` before calling it, and prefer a narrow date range.

        The precondition is re-checked server-side, because every one of these jobs reports success while
        doing nothing when it is unmet.
      parameters:
        - name: job
          in: path
          required: true
          schema:
            type: string
            enum:
              - noise-bootstrap
              - noise-sync
              - wbgt-fill
              - wbgt-scrape
              - wbgt-water-parade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectCode:
                  type: string
                startDate:
                  type: string
                  format: date
                  description: YYYY-MM-DD, inclusive.
                endDate:
                  type: string
                  format: date
                  description: YYYY-MM-DD, inclusive.
                flags:
                  type: object
                  additionalProperties:
                    type: boolean
                  description: Only flags the job declares are forwarded.
              required:
                - projectCode
                - startDate
                - endDate
      responses:
        "200":
          description: The service's own response, passed through.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400": *a3
        "401": *a1
        "403": *a2
        "404": *a4
        "503":
          description: The service's base URL is not configured on HALO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        A token minted by `scripts/mint-api-token.mjs`, sent as `Authorization: Bearer halo_…`.

        Scopes are additive and **not** hierarchical: `read` permits reads, `write` permits configuration
        changes, `jobs` permits triggering jobs. Holding `write` does not confer `read`, and neither confers
        `jobs`. Call `getSession` to see what a token actually carries.

        Every change a token makes is attributed to `agent:<name>` in the audit trail.
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem detail. Branch on `title`, not on prose.
      properties:
        error:
          type: string
          description: Human-readable explanation of what went wrong.
      required:
        - error
...